+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/advance1/Advance1Activity.java
@@ -14,25 +14,23 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.advance1;
+package im.ene.lab.toro.sample.presentation.advance1;
 
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
 
 /**
- * Created by eneim on 6/29/16.
- *
- * A sample {@link RecyclerView.ViewHolder} skeleton.
+ * Created by eneim on 6/30/16.
  */
-public abstract class Advance1ViewHolder extends RecyclerView.ViewHolder {
+public class Advance1Activity extends AppCompatActivity {
 
-  static int TYPE_VIDEO = 1;
-
-  static int TYPE_NORMAL = 2;
-
-  public Advance1ViewHolder(View itemView) {
-    super(itemView);
+  @Override protected void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    if (savedInstanceState == null) {
+      getSupportFragmentManager().beginTransaction()
+          .replace(android.R.id.content, Advance1ListFragment.newInstance())
+          .commit();
+    }
   }
-
-  public abstract void bind(RecyclerView.Adapter adapter, Object item);
 }
