+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/advance1/Advance1BaseVideoViewHolder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.advance1;
+package im.ene.lab.toro.sample.presentation.advance1;
 
 import android.view.View;
 import im.ene.lab.toro.PlayerViewHelper;
